requirements.txt
opencv-python==4.11.0.86
mediapipe==0.10.20
numpy==1.26.0
supabase==2.12.0
python-dotenv==1.0.1
Flask==3.1.0
Flask-Cors==5.0.0
Pillow==11.1.0